Removal of codeBundle
I removed codebundle because i will add an option for the admin to send a code directly using an email.
thus there will be a code issued on the email entered by the admin and i need to figure out how to manage all this.
also i need to see what to do for the authorisations and the roles
Need to have a better look on SDZ tuto for the userBundle without fosuserbundle
I need also to look at firewalls and security params